api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.31.1
groq>=0.12.0
httpx>=0.28.1
python-dotenv>=1.1.0
pydantic>=2.11.0
